Remove paymentData from state and intent and directly trigger onSubmit after handling the google pay SDK result
COSDK-1310
