+++ googlepay/src/main/java/com/adyen/checkout/googlepay/internal/ui/state/GooglePayComponentState.kt
@@ -10,7 +10,6 @@ package com.adyen.checkout.googlepay.internal.ui.state
 
 import com.adyen.checkout.core.components.internal.ui.state.ComponentState
 import com.adyen.checkout.googlepay.GooglePayButtonStyling
-import com.google.android.gms.wallet.PaymentData
 
 internal data class GooglePayComponentState(
     val allowedPaymentMethods: String,
@@ -18,5 +17,4 @@ internal data class GooglePayComponentState(
     val isButtonVisible: Boolean,
     val isLoading: Boolean,
     val isAvailable: Boolean,
-    val paymentData: PaymentData?,
 ) : ComponentState
