+++ googlepay/src/main/java/com/adyen/checkout/googlepay/internal/ui/state/GooglePayComponentStateFactory.kt
@@ -22,6 +22,5 @@ internal class GooglePayComponentStateFactory(
         isButtonVisible = true,
         isLoading = false,
         isAvailable = false,
-        paymentData = null,
     )
 }
