+++ googlepay/src/main/java/com/adyen/checkout/googlepay/internal/ui/state/GooglePayIntent.kt
@@ -9,15 +9,12 @@
 package com.adyen.checkout.googlepay.internal.ui.state
 
 import com.adyen.checkout.core.components.internal.ui.state.ComponentStateIntent
-import com.google.android.gms.wallet.PaymentData
 
 internal sealed interface GooglePayIntent : ComponentStateIntent {
 
     data class UpdateLoading(val isLoading: Boolean) : GooglePayIntent
 
     data class UpdateButtonVisible(val isButtonVisible: Boolean) : GooglePayIntent
-
-    data class UpdatePaymentData(val paymentData: PaymentData?) : GooglePayIntent
 
     data class UpdateAvailability(val isAvailable: Boolean) : GooglePayIntent
 }
