+++ googlepay/src/test/java/com/adyen/checkout/googlepay/internal/ui/state/GooglePayComponentStateReducerTest.kt
@@ -8,22 +8,16 @@
 
 package com.adyen.checkout.googlepay.internal.ui.state
 
-import com.google.android.gms.wallet.PaymentData
-import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
-import org.mockito.Mock
 import org.mockito.junit.jupiter.MockitoExtension
 
 @ExtendWith(MockitoExtension::class)
 internal class GooglePayComponentStateReducerTest {
 
     private lateinit var reducer: GooglePayComponentStateReducer
-
-    @Mock
-    private lateinit var paymentData: PaymentData
 
     @BeforeEach
     fun beforeEach() {
@@ -49,15 +43,6 @@ internal class GooglePayComponentStateReducerTest {
     }
 
     @Test
-    fun `when intent is UpdatePaymentData, then paymentData is updated`() {
-        val state = createInitialState()
-
-        val actual = reducer.reduce(state, GooglePayIntent.UpdatePaymentData(paymentData))
-
-        assertEquals(paymentData, actual.paymentData)
-    }
-
-    @Test
     fun `when intent is UpdateAvailability, then isAvailable is updated`() {
         val state = createInitialState()
 
@@ -72,6 +57,5 @@ internal class GooglePayComponentStateReducerTest {
         isButtonVisible = false,
         isLoading = false,
         isAvailable = false,
-        paymentData = null,
     )
 }
